Split font interface
